pyproject.toml
[tool.poetry]
name = "analytics-publisher"
version = "0.1.0"
description = ""
authors = ["Dddarknight <9801677@gmail.com>"]
readme = "README.md"
packages = [{include = "analytics_publisher"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
sqlalchemy = "^1.4.44"
motor = "^3.1.1"
python-dotenv = "^0.21.0"
uvicorn = "^0.20.0"
celery = "^5.2.7"
pandas = "^1.5.2"
pydrive = "^1.3.1"
asyncio = "^3.4.3"
seaborn = "^0.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
